# swagger/swagger.yaml
openapi: 3.0.0
...
info:
  title: swagger
  version: '1.0'
  description: ''
host: 'localhost:3000'
schemes:
  - http
produces:
  - application/json
consumes:
  - application/json
paths:
...
  /games/:
    post:
      summary: Create Game
      operationId: post-games
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Game'
      description: 新しくゲームを作成する。
      tags:
        - Game
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                roomId:
                  type: string
...
  '/games/{gameId}/players':
    parameters:
      - schema:
          type: string
        name: gameId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags:
        - Game
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  players:
                    type: array
                    items:
                      $ref: '#/components/schemas/GamePlayer'
      operationId: get-games-gameId-users
      security:
        - authorization: []
  '/games/{gameId}/players/me':
    parameters:
      - schema:
          type: string
        name: gameId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags: []
      responses:
        '200':
          description: OK
          headers: {}
          content:
            application/json:
              schema:
                type: object
                properties:
                  player:
                    $ref: '#/components/schemas/GamePlayer'
      operationId: get-games-gameId-players-me
      security:
        - authorization: []
  '/games/{gameId}/players/{playerId}':
    parameters:
      - schema:
          type: string
        name: gameId
        in: path
        required: true
      - schema:
          type: string
        name: playerId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      tags:
        - Game
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  player:
                    $ref: '#/components/schemas/GamePlayer'
      operationId: get-games-gameId-players-playerId
      security:
        - authorization: []
...
securityDefinitions:
  Authorization:
    name: Authorization
    type: apiKey
    in: header
servers:
  - description: ''
    url: 'http://localhost:3000/'
components:
  securitySchemes:
    authorization:
      type: http
      scheme: bearer
      description: ''
...
  schemas:
...
    GamePlayer:
      title: GamePlayer
      x-stoplight:
        id: nwdk5sktjysqx
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        gameId:
          type: string
        isFramer:
          type: boolean
        character:
          type: object
          properties:
            id:
              type: string
            displayName:
              type: string
            imageUrl:
              type: string
...
    Game:
      title: Game
      x-stoplight:
        id: drxq2nvw3khjb
      type: object
      properties:
        id:
          type: string
        room_id:
          type: string
        status:
          type: object
          properties:
            id:
              type: string
            code:
              type: string